backend/bot/requirements.txt
python-telegram-bot==21.0.1
requests
cloudinary
firebase-admin
python-dotenv
pydantic
